Remove things that are not being used.
I'm not sure if this is useful or not. There two things confused me for
a little while. I'm pretty sure they're perhaps legacy code or
redundant. When I deleted them, it broke no tests so deleting them here
but please feel free to close this PR if I'm missing something and they
have a purpose.