Add an acceptance criteria.
When using the heuristic to optimise the objective function it was doing
so in ignorance of the constraints.

This add the ability to pass a `acceptance_criteria` to both heuristics
which gives a bound on some measure. In practice it's used in
`schedule.heuristic` when an objective function is passed. In this case
it passes the violation counts as an `acceptance_criteria` to ensure
that as each heuristic goes through the search space optimising the
objective function it will only accept a solution if this does not add
any constraint violations.

TLDR:

The modified tests show a good example of this. The objective function
optimisation was giving a schedule that was in fact invalid. This has
now been fixed.